fix filtering on status and make sure status is always set
changes the filtering logic a bit for statusses. We now only make statusses
optional if we aren't filtering on them. For displaying unknown
statusses (either empty string or not set) we now use the status 'Onbekend'.

Also made the status_color check in the detail a bit easier to read and easier
to extend later on by using a switch statement